requirements.txt
chardet
tabulate
jupyter
sqlalchemy
pycryptodome